# backend/api-doc.yaml
openapi: 3.0.0
info:
  title: Kulhad Flask Backend API
  version: 1.0.0
  description: |
    OpenAPI 3.0 specification generated from the current Flask backend blueprints.

    Authentication-protected endpoints use JWT bearer authentication.
    The computer vision endpoints upload an image to the backend, which then calls an external object-detection service internally.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/orders/checkout-profile:
    get:
      tags:
        - orders
      summary: Get checkout profile
      description: Returns the authenticated customer's checkout address profile.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Checkout profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutProfile'
              example:
                name: Customer User
                email: customer@example.com
                phone: "9999999999"
                address: Main Street
                city: Jaipur
                state: Rajasthan
                postal_code: "302001"
                has_saved_address: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-customer account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Checkout profile is only available for customers
        '404':
          $ref: '#/components/responses/UserNotFound'
    put:
      tags:
        - orders
      summary: Update checkout profile
      description: Updates the authenticated customer's saved checkout address.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutProfileUpdateRequest'
            example:
              phone: "9999999999"
              address: Main Street
              city: Jaipur
              state: Rajasthan
              postal_code: "302001"
      responses:
        '200':
          description: Checkout profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-customer account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Checkout profile is only available for customers
        '404':
          $ref: '#/components/responses/UserNotFound'
  /api/orders/:
    post:
      tags:
        - orders
      summary: Place an order
      description: Creates a customer order, decrements stock, calculates shipping, and creates a payment record.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
            example:
              phone: "9999999999"
              address: Main Street
              city: Jaipur
              state: Rajasthan
              postal_code: "302001"
              payment_method: UPI
              items:
                - id: 1
                  qty: 2
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreateResponse'
              example:
                message: Order placed successfully
                id: 12
                order_number: ORD-0012
                status: Pending
                payment: Paid
                method: UPI
                amount: 92.0
                items:
                  - name: Classic Kulhad
                    quantity: 2
                    subtotal: 42.0
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-customer account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Only customer accounts can place orders
        '404':
          description: User or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user:
                  value:
                    error: User not found
                product:
                  value:
                    error: One or more products could not be found
  /api/orders/admin:
    get:
      tags:
        - orders
      summary: List orders for admins
      description: Returns all orders enriched with customer, payment, and item information.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/orders/my-history:
    get:
      tags:
        - orders
      summary: Get customer order history
      description: Returns order history for the authenticated customer.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-customer account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Order history is only available for customers
        '404':
          $ref: '#/components/responses/UserNotFound'
  /api/orders/{order_id}/status:
    put:
      tags:
        - orders
      summary: Update order and payment status
      description: Updates an order status and optionally its payment status. Admin access is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdateRequest'
            example:
              status: shipped
              payment: paid
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusUpdateResponse'
              example:
                message: Order updated
                id: 12
                status: Shipped
                payment: Paid
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
  /api/orders/bulk-payment-status:
    put:
      tags:
        - orders
      summary: Bulk update payment status
      description: Updates payment status for multiple orders. Admin access is required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkPaymentStatusRequest'
            example:
              order_ids:
                - 1
                - 2
              payment: paid
      responses:
        '200':
          description: Payment statuses updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPaymentStatusResponse'
              example:
                message: Updated payment status for 2 order(s)
                orders:
                  - id: 1
                    payment: Paid
                  - id: 2
                    payment: Paid
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: integer
      description: Order identifier.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: Authenticated user does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Access forbidden
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/MessageResponse'
    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: User not found
...
    OrderNotFound:
      description: Order not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Order not found
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthErrorResponse:
      type: object
      properties:
        msg:
          type: string
      required:
        - msg
...
    CheckoutProfile:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        has_saved_address:
          type: boolean
      required:
        - name
        - email
        - phone
        - address
        - city
        - state
        - postal_code
        - has_saved_address
    CheckoutProfileUpdateRequest:
      type: object
      properties:
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
    OrderLineItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: integer
        subtotal:
          type: number
          format: float
      required:
        - name
        - quantity
        - subtotal
    OrderHistoryItem:
      allOf:
        - $ref: '#/components/schemas/OrderLineItem'
        - type: object
          properties:
            id:
              type: integer
            price:
              type: number
              format: float
            stock:
              type: integer
            image:
              type: string
    Order:
      type: object
      properties:
        id:
          type: integer
        order_number:
          type: string
        status:
          type: string
        payment:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        method:
          type: string
        amount:
          type: number
          format: float
        date:
          type: string
          format: date
        items:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/OrderLineItem'
              - $ref: '#/components/schemas/OrderHistoryItem'
        item_count:
          type: integer
      required:
        - id
        - order_number
        - status
        - payment
        - name
        - email
        - phone
        - address
        - method
        - amount
        - date
        - items
        - item_count
...
    OrderCreateItem:
      type: object
      required:
        - id
        - qty
      properties:
        id:
          type: integer
        qty:
          type: integer
    OrderCreateRequest:
      type: object
      required:
        - items
      properties:
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        payment_method:
          type: string
          enum:
            - Cash on Delivery
            - UPI
            - Credit / Debit Card
        shipping:
          type: number
          format: float
          description: Client-sent shipping is validated only for non-negative input; final shipping is determined server-side.
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderCreateItem'
    OrderCreateResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: integer
        order_number:
          type: string
        status:
          type: string
        payment:
          type: string
        method:
          type: string
        amount:
          type: number
          format: float
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineItem'
      required:
        - message
        - id
        - order_number
        - status
        - payment
        - method
        - amount
        - items
    OrderStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - confirmed
            - shipped
            - delivered
            - cancelled
        payment:
          type: string
          enum:
            - pending
            - paid
            - failed
            - refunded
    OrderStatusUpdateResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: integer
        status:
          type: string
        payment:
          type: string
      required:
        - message
        - id
        - status
        - payment
    BulkPaymentStatusRequest:
      type: object
      required:
        - order_ids
        - payment
      properties:
        order_ids:
          type: array
          items:
            type: integer
        payment:
          type: string
          enum:
            - pending
            - paid
            - failed
            - refunded
    BulkPaymentStatusResponse:
      type: object
      properties:
        message:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              payment:
                type: string
            required:
              - id
              - payment
      required:
        - message
        - orders
...